+++ packages/block-library/src/media-text/edit.js
@@ -32,7 +32,7 @@ import {
 } from '@wordpress/components';
 import { isBlobURL, getBlobTypeByURL } from '@wordpress/blob';
 import { pullLeft, pullRight } from '@wordpress/icons';
-import { store as coreStore } from '@wordpress/core-data';
+import { useEntityProp, store as coreStore } from '@wordpress/core-data';
 
 /**
  * Internal dependencies
@@ -132,7 +132,12 @@ function attributesFromMedia( {
 	};
 }
 
-function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
+function MediaTextEdit( {
+	attributes,
+	isSelected,
+	setAttributes,
+	context: { postId, postType },
+} ) {
 	const {
 		focalPoint,
 		href,
@@ -150,8 +155,41 @@ function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
 		rel,
 		verticalAlignment,
 		allowedBlocks,
+		useFeaturedImage,
 	} = attributes;
 	const mediaSizeSlug = attributes.mediaSizeSlug || DEFAULT_MEDIA_SIZE_SLUG;
+
+	const [ featuredImage ] = useEntityProp(
+		'postType',
+		postType,
+		'featured_media',
+		postId
+	);
+
+	const featuredImageMedia = useSelect(
+		( select ) =>
+			featuredImage &&
+			select( coreStore ).getMedia( featuredImage, { context: 'view' } ),
+		[ featuredImage ]
+	);
+
+	const featuredImageURL = useFeaturedImage
+		? featuredImageMedia?.source_url
+		: '';
+	const featuredImageAlt = useFeaturedImage
+		? featuredImageMedia?.alt_text
+		: '';
+
+	const toggleUseFeaturedImage = () => {
+		setAttributes( {
+			imageFill: false,
+			mediaType: 'image',
+			mediaId: undefined,
+			mediaUrl: undefined,
+			mediaAlt: undefined,
+			useFeaturedImage: ! useFeaturedImage,
+		} );
+	};
 
 	const { imageSizes, image } = useSelect(
 		( select ) => {
@@ -294,7 +332,9 @@ function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
 					/>
 				</ToolsPanelItem>
 			) }
-			{ imageFill && mediaUrl && mediaType === 'image' && (
+			{ imageFill &&
+				( mediaUrl || featuredImageURL ) &&
+				mediaType === 'image' && (
 				<ToolsPanelItem
 					label={ __( 'Focal point' ) }
 					isShownByDefault
@@ -307,7 +347,11 @@ function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
 						__nextHasNoMarginBottom
 						__next40pxDefaultSize
 						label={ __( 'Focal point' ) }
-						url={ mediaUrl }
+						url={
+							useFeaturedImage && featuredImageURL
+								? featuredImageURL
+								: mediaUrl
+						}
 						value={ focalPoint }
 						onChange={ ( value ) =>
 							setAttributes( { focalPoint: value } )
@@ -317,7 +361,7 @@ function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
 					/>
 				</ToolsPanelItem>
 			) }
-			{ mediaType === 'image' && (
+			{ mediaType === 'image' && ( mediaUrl || featuredImageURL ) && (
 				<ToolsPanelItem
 					label={ __( 'Alternative text' ) }
 					isShownByDefault
@@ -327,7 +371,7 @@ function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
 					<TextareaControl
 						__nextHasNoMarginBottom
 						label={ __( 'Alternative text' ) }
-						value={ mediaAlt }
+						value={ mediaAlt || featuredImageAlt }
 						onChange={ onMediaAltChange }
 						help={
 							<>
@@ -411,7 +455,11 @@ function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
 						onChangeUrl={ onSetHref }
 						linkDestination={ linkDestination }
 						mediaType={ mediaType }
-						mediaUrl={ image && image.source_url }
+						mediaUrl={
+							useFeaturedImage && featuredImageURL
+								? featuredImageURL
+								: image && image.source_url
+						}
 						mediaLink={ image && image.link }
 						linkTarget={ linkTarget }
 						linkClass={ linkClass }
@@ -428,6 +476,7 @@ function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
 					commitWidthChange={ commitWidthChange }
 					ref={ refMediaContainer }
 					enableResize={ blockEditingMode === 'default' }
+					toggleUseFeaturedImage={ toggleUseFeaturedImage }
 					{ ...{
 						focalPoint,
 						imageFill,
@@ -439,6 +488,9 @@ function MediaTextEdit( { attributes, isSelected, setAttributes } ) {
 						mediaType,
 						mediaUrl,
 						mediaWidth,
+						useFeaturedImage,
+						featuredImageURL,
+						featuredImageAlt,
 					} }
 				/>
 				{ mediaPosition !== 'right' && <div { ...innerBlocksProps } /> }