+++ packages/react-native-bridge/index.js
@@ -8,6 +8,11 @@ import { NativeModules, NativeEventEmitter, Platform } from 'react-native';
  */
 import RCTAztecView from '@wordpress/react-native-aztec';
 
+/**
+ * Internal dependencies
+ */
+import parseException from './lib/parseException';
+
 const { RNReactNativeGutenbergBridge } = NativeModules;
 const isIOS = Platform.OS === 'ios';
 const isAndroid = Platform.OS === 'android';
@@ -444,6 +449,62 @@ export function toggleUndoButton( isDisabled ) {
 
 export function toggleRedoButton( isDisabled ) {
 	RNReactNativeGutenbergBridge.toggleRedoButton( isDisabled );
+}
+
+/**
+ * Log exception to host app's crash logging service.
+ * @param {Object}   exception         Exception object
+ * @param {Object}   [extra]           Extra parameters to include in the exception.
+ * @param {Object}   [extra.context]   Context of the exception.
+ * @param {Object}   [extra.tags]      Tags to associate with the exception.
+ * @param {Object}   [extra.isHandled] True if the exception is handled.
+ * @param {Object}   [extra.handledBy] The mechanism that detected the exception.
+ * @param {Function} [callback]        Callback triggered when the exception is sent.
+ */
+export function logException(
+	exception,
+	{ context, tags, isHandled, handledBy } = {
+		context: {},
+		tags: {},
+		isHandled: false,
+		handledBy: 'Unknown',
+	},
+	callback
+) {
+	const parsedException = {
+		...parseException( exception, { context, tags } ),
+		isHandled,
+		handledBy,
+	};
+
+	const onLogException = ( wasSent ) => {
+		if ( ! wasSent ) {
+			// eslint-disable-next-line no-console
+			console.error(
+				'An error ocurred when logging the exception',
+				parsedException
+			);
+		}
+		if ( callback ) {
+			callback();
+		}
+	};
+
+	// Only log exceptions in production
+	// eslint-disable-next-line no-undef
+	if ( __DEV__ ) {
+		// eslint-disable-next-line no-console
+		console.info( 'Exception that would be logged', parsedException );
+		if ( callback ) {
+			callback();
+		}
+		return;
+	}
+
+	RNReactNativeGutenbergBridge.logException(
+		parsedException,
+		onLogException
+	);
 }
 
 export default RNReactNativeGutenbergBridge;