+++ packages/react-native-editor/CHANGELOG.md
@@ -13,6 +13,10 @@ For each user feature we should also add a importance categorization label  to i
 -   [*] Improve consistency of the block outline indicating the currently selected block [#59415]
 -   [*] Gallery block items with in-progress, paused, or failed media uploads correctly display an highlight outline when selected [#59423]
 -   [**] [internal] Upgrade React Native to version 0.73.3 [#58475]
+-   [**] Add error boundary components and exception logging [#59221]
+
+## 1.114.1
+-   [**] Fix a crash produced when the content of a synced pattern is updated [#59632]
 
 ## 1.114.0
 -   [*] Prevent crash when autoscrolling to blocks [#59110]